requirements.txt
python-dotenv
tenacity
feedparser
requests
beautifulsoup4
lxml
Pillow
